requirements.txt
# 3x Leveraged Sentiment-Driven Trading System
# Python dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data Scraping & Parsing
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
feedparser>=6.0.10

# Financial Data
yfinance>=0.2.30
pandas>=2.1.0
numpy>=1.24.0

# LLM Integration
ollama>=0.1.7
python-dotenv>=1.0.0

# Backtesting
vectorbt>=0.24.0

# Utilities
aiohttp>=3.9.0
asyncio>=3.4.3

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
